reshape api vehicle occupancies
