+++ api_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/CanalTP/forseti/internal/freefloatings"
 	"github.com/CanalTP/forseti/internal/parkings"
 	"github.com/CanalTP/forseti/internal/utils"
+	"github.com/CanalTP/forseti/internal/vehicleoccupancies"
 )
 
 func TestStatusApiExist(t *testing.T) {
@@ -117,22 +118,24 @@ func TestVehicleOccupanciesAPIWithDataFromFile(t *testing.T) {
 	loc, err := time.LoadLocation("Europe/Paris")
 	require.Nil(err)
 
+	vehiculeOccupanciesContext := &vehicleoccupancies.VehicleOccupanciesContext{}
+
 	// Load StopPoints from file .../mapping_stops.csv
 	uri, err := url.Parse(fmt.Sprintf("file://%s/", fixtureDir))
 	require.Nil(err)
-	stopPoints, err := LoadStopPoints(*uri, defaultTimeout)
+	stopPoints, err := vehicleoccupancies.LoadStopPoints(*uri, defaultTimeout)
 	require.Nil(err)
-	manager.InitStopPoint(stopPoints)
-	assert.Equal(len(*manager.stopPoints), 25)
+	vehiculeOccupanciesContext.InitStopPoint(stopPoints)
+	assert.Equal(len(vehiculeOccupanciesContext.GetStopPoints()), 25)
 
 	// Load courses from file .../extraction_courses.csv
 	uri, err = url.Parse(fmt.Sprintf("file://%s/", fixtureDir))
 	require.Nil(err)
-	courses, err := LoadCourses(*uri, defaultTimeout)
+	courses, err := vehicleoccupancies.LoadCourses(*uri, defaultTimeout)
 	require.Nil(err)
-	manager.InitCourse(courses)
-	assert.Equal(len(*manager.courses), 1)
-	coursesFor40 := (*manager.courses)["40"]
+	vehiculeOccupanciesContext.InitCourse(courses)
+	assert.Equal(len(vehiculeOccupanciesContext.GetCourses()), 1)
+	coursesFor40 := (vehiculeOccupanciesContext.GetCourses())["40"]
 	assert.Equal(len(coursesFor40), 310)
 
 	// Load RouteSchedules from file
@@ -149,9 +152,9 @@ func TestVehicleOccupanciesAPIWithDataFromFile(t *testing.T) {
 	require.Nil(err)
 	sens := 0
 	startIndex := 1
-	routeSchedules := LoadRouteSchedulesData(startIndex, navitiaRoutes, sens, loc)
-	manager.InitRouteSchedule(routeSchedules)
-	assert.Equal(len(*manager.routeSchedules), 141)
+	routeSchedules := vehicleoccupancies.LoadRouteSchedulesData(startIndex, navitiaRoutes, sens, loc)
+	vehiculeOccupanciesContext.InitRouteSchedule(routeSchedules)
+	assert.Equal(len(vehiculeOccupanciesContext.GetRouteSchedules()), 141)
 
 	// Load prediction from a file
 	uri, err = url.Parse(fmt.Sprintf("file://%s/predictions.json", fixtureDir))
@@ -165,18 +168,19 @@ func TestVehicleOccupanciesAPIWithDataFromFile(t *testing.T) {
 	predicts := &data.PredictionData{}
 	err = json.Unmarshal([]byte(jsonData), predicts)
 	require.Nil(err)
-	predictions := LoadPredictionsData(predicts, loc)
+	predictions := vehicleoccupancies.LoadPredictionsData(predicts, loc)
 	assert.Equal(len(predictions), 65)
 
-	occupanciesWithCharge := CreateOccupanciesFromPredictions(&manager, predictions)
-	manager.UpdateVehicleOccupancies(occupanciesWithCharge)
-	assert.Equal(len(*manager.vehicleOccupancies), 35)
+	occupanciesWithCharge := vehicleoccupancies.CreateOccupanciesFromPredictions(vehiculeOccupanciesContext, predictions)
+	vehiculeOccupanciesContext.UpdateVehicleOccupancies(occupanciesWithCharge)
+	assert.Equal(len(vehiculeOccupanciesContext.GetVehiclesOccupancies()), 35)
 
 	c, engine := gin.CreateTestContext(httptest.NewRecorder())
+	vehicleoccupancies.AddVehicleOccupanciesEntryPoint(engine, vehiculeOccupanciesContext)
 	engine = SetupRouter(&manager, engine)
 
 	// Request without any parameter (Date with default value = Now().Format("20060102"))
-	response := VehicleOccupanciesResponse{}
+	response := vehicleoccupancies.VehicleOccupanciesResponse{}
 	c.Request = httptest.NewRequest("GET", "/vehicle_occupancies", nil)
 	w := httptest.NewRecorder()
 	engine.ServeHTTP(w, c.Request)
@@ -197,7 +201,7 @@ func TestVehicleOccupanciesAPIWithDataFromFile(t *testing.T) {
 	assert.Len(response.VehicleOccupancies, 35)
 	assert.Empty(response.Error)
 
-	resp := VehicleOccupanciesResponse{}
+	resp := vehicleoccupancies.VehicleOccupanciesResponse{}
 	c.Request = httptest.NewRequest(
 		"GET",
 		"/vehicle_occupancies?date=20210118&vehiclejourney_id=vehicle_journey:0:123713792-1",
@@ -211,7 +215,7 @@ func TestVehicleOccupanciesAPIWithDataFromFile(t *testing.T) {
 	assert.Len(resp.VehicleOccupancies, 7)
 	assert.Empty(resp.Error)
 
-	resp = VehicleOccupanciesResponse{}
+	resp = vehicleoccupancies.VehicleOccupanciesResponse{}
 	c.Request = httptest.NewRequest("GET",
 		"/vehicle_occupancies?date=20210118&vehiclejourney_id=vehicle_journey:0:123713792-1&stop_id=stop_point:0:SP:80:4121",
 		nil)
@@ -232,6 +236,7 @@ func TestVehicleOccupanciesAPIWithDataFromFile(t *testing.T) {
 	assert.Equal(resp.VehicleOccupancies[0].Occupancy, 11)
 
 	// Verify /status
+	manager.SetVehiculeOccupanciesContext(vehiculeOccupanciesContext)
 	c.Request = httptest.NewRequest("GET", "/status", nil)
 	w = httptest.NewRecorder()
 	engine.ServeHTTP(w, c.Request)
